field_specials through sub_80CBFA0

+++ include/field_specials.h
@@ -12,7 +12,7 @@ enum HiddenItemAttr
 };
 
 u8 GetLeadMonIndex(void);
-u8 CountDigits(u16 number);
+s32 CountDigits(s32 number);
 void TV_PrintIntToStringVar(u8, int);
 u16 ScrSpecial_GetStarter(void);
 void sub_80CCB68(void);
